+++ nodeconductor/cost_tracking/migrations/0018_priceestimate_threshold.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+import django.core.validators
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('cost_tracking', '0017_nullable_object_id'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='priceestimate',
+            name='threshold',
+            field=models.FloatField(default=0, validators=[django.core.validators.MinValueValidator(0)]),
+        ),
+    ]